+++ src/main/java/com/example/crossmintchallenge/gateway/PolyanetsGateway.java
@@ -3,7 +3,7 @@ package com.example.crossmintchallenge.gateway;
 import com.example.crossmintchallenge.gateway.dto.PolyanetResponse;
 
 public interface PolyanetsGateway {
-    PolyanetResponse postPolyanet(Long row, Long column);
+    PolyanetResponse postPolyanet(int row, int column);
 
-    void deletePolyanet(Long row, Long column);
+    void deletePolyanet(int row, int column);
 }
